backend/requirements.txt
flask
flask_sqlalchemy
flask_cors
psycopg2-binary
flask_bcrypt
flask_jwt_extended
flask_migrate
openai
transformers 
torch 
scikit_learn
pillow
SpeechRecognition
spacy
pydub
soundfile 
numpy



#python -m spacy download en_core_web_sm
#python -m spacy download es_core_news_sm


#react-icons

#instalar en el frontend 
#npm install bootstrap
#npm install react-spinners
#npm install recorder-js
#npm install audiorecorder
#npm install framer-motion